// src/api/auth-service/models/api-response-authenticate-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AuthenticateResponse } from './authenticate-response';

/**
 * 
 * @export
 * @interface ApiResponseAuthenticateResponse
 */
export interface ApiResponseAuthenticateResponse {
    /**
     * 
     * @type {number}
     * @memberof ApiResponseAuthenticateResponse
     */
    'code'?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiResponseAuthenticateResponse
     */
    'message'?: string;
    /**
     * 
     * @type {AuthenticateResponse}
     * @memberof ApiResponseAuthenticateResponse
     */
    'result'?: AuthenticateResponse;
}