// src/api/auth-service/models/api-response-introspect-token-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { IntrospectTokenResponse } from './introspect-token-response';

/**
 * 
 * @export
 * @interface ApiResponseIntrospectTokenResponse
 */
export interface ApiResponseIntrospectTokenResponse {
    /**
     * 
     * @type {number}
     * @memberof ApiResponseIntrospectTokenResponse
     */
    'code'?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiResponseIntrospectTokenResponse
     */
    'message'?: string;
    /**
     * 
     * @type {IntrospectTokenResponse}
     * @memberof ApiResponseIntrospectTokenResponse
     */
    'result'?: IntrospectTokenResponse;
}